Refactor operate into `mod`
